requirements.txt
django==4.1.4
django-environ==0.9.0
django-debug-toolbar==3.8.1

attrs==22.1.0

psycopg2-binary==2.9.5
pillow==9.3.0

django-filter==22.1

django-leaflet==0.28.3
django-bootstrap5==22.2

gunicorn

mypy==0.991
django-stubs==1.13.1
